refactor:  diary 페이지 코드 리펙토링 ( 한페이지로 )
